Added updates to mythic panel for changing how state works and fixing bug with shared state across users

+++ modules/UI/mythicplus/mythicplus.server.ts
@@ -1,7 +1,14 @@
 /** @ts-expect-error */
 let aio: AIO = {}; 
 
-function getPlayerGroupId(player: Player): number {
+import { Logger } from "../../classes/logger";
+import type { MythicPlusState } from "./mythicplus.state";
+
+const logger = new Logger("MythicPlusMod");
+const StateStorage: Map<number, MythicPlusState> = new Map();
+
+// This looks up the current group id for the player -1 indicates no group
+function getPlayerGroupId(player: Player): number {    
     const result = CharDBQuery(`SELECT m.guid FROM acore_characters.characters c left join acore_characters.group_member m on c.guid = m.memberGuid where c.guid = ${player.GetGUID()}`);
 
     if(!result) {
@@ -11,26 +18,23 @@ function getPlayerGroupId(player: Player): number {
     return result.GetUInt32(0);
 }
 
-
 // Get the difficulty alread set for the player or group
 function _getDifficulty(player: Player): number {
     const difficulty = player.GetDifficulty(); 
-    const group = player.GetGroup();
-    if(!group) {
+    const groupId = getPlayerGroupId(player);
+
+
+    logger.debug(`MythicPlusMod: Getting difficulty for ${player.GetName()} with difficulty ${difficulty} and group ${groupId}`);
+
+    if(groupId == -1) {
         aio.Handle(player, "MythicPlus", "SetDifficulty", difficulty);
     }
 
-    const result = CharDBQuery(`SELECT difficulty FROM group_difficulty WHERE guid = ${group.GetGUID()}`);
+    const result = CharDBQuery(`SELECT difficulty FROM group_difficulty WHERE guid = ${groupId}`);
     if(result) {
-        print(`MythicPlusMod: Setting difficulty for ${player.GetName()} to ${result.GetUInt32(0)}`);
-        aio.Handle(player, 'MythicPlus', 'SetDifficulty', result.GetUInt32(0));        
-        return;
+        logger.debug(`MythicPlusMod: Setting difficulty for ${player.GetName()} to ${result.GetUInt32(0)}`);             
+        return result.GetUInt32(0);
     } 
-}
-
-// Server API for Client
-function GetDifficulty(this:void, player: Player): void {
-    aio.Handle(player, 'MythicPlus', 'SetDifficulty', _getDifficulty(player));
 }
 
 // Set the difficulty for the encounter
@@ -41,19 +45,24 @@ function _setDifficulty(player: Player, difficulty: number): void {
         player.SendNotification('You must be in a group to set a mythic+ difficulty');        
         return;
     }
-    PrintInfo("Group ID: " + groupId);    
+    logger.debug(`Setting difficulty for ${player.GetName()} to ${difficulty}`);    
 
     if(! group.IsLeader(player.GetGUID())) {
         return;
     }
 
+    const map = player.GetMap();
+    if(map.IsDungeon() != false) {
+        player.SendNotification('You can not change the difficulty in a dungeon');        
+        return;
+    }
+
     if(difficulty > 4) {
-        PrintError(`MythicPlusMod: Invalid difficulty set:  ${difficulty}`);
+        logger.error(`Invalid difficulty set:  ${difficulty}`);
     }
   
     if(difficulty == 0) {
-        CharDBExecute(`DELETE FROM group_difficulty WHERE guid = ${groupId}`);
-        GetDifficulty(player);        
+        CharDBExecute(`DELETE FROM group_difficulty WHERE guid = ${groupId}`);        
         return;
     }
 
@@ -62,22 +71,56 @@ function _setDifficulty(player: Player, difficulty: number): void {
 
 function SetDifficulty(this:void, player: Player, difficulty: number): void {
     _setDifficulty(player, difficulty);    
+    aio.Handle(player, 'MythicPlus', 'UpdateState', StateStorage.get(player.GetGUIDLow()));     
 }
 
-const ShowIt: player_event_on_command = (event: number,player: Player, command: string): boolean => {
-    if(command == 'mythicplus') {        
-        const difficulty = player.GetDifficulty();
-        PrintInfo(`MythicPlusMod: Showing UI for ${player.GetName()} with difficulty ${difficulty}`);
-        aio.Handle(player, 'MythicPlus', 'ShowUI', {difficulty: difficulty}); 
+function _refreshState(player: Player) {
+    if(player.IsInGroup()) {
+        const groupId = getPlayerGroupId(player);
+        const groupLeader = player.GetGroup().GetLeaderGUID();
+        const isLeader = player.GetGUID() == groupLeader;
+        const difficulty = _getDifficulty(player);
+        StateStorage.set(player.GetGUIDLow(), {difficulty, inGroup: true, groupId, groupLeader, isLeader});
+        return;
+    } else {
+        StateStorage.set(player.GetGUIDLow(), {difficulty: _getDifficulty(player), inGroup: false, groupId: -1, groupLeader: -1, isLeader: false});        
+    }
+}
+
+// Update the state from what is on the server and send it back to the client. 
+function GetState(this:void, player: Player): void {
+    _refreshState(player);
+    const state = StateStorage.get(player.GetGUIDLow());
+    aio.Handle(player, 'MythicPlus', 'UpdateState', state);
+}
+
+const OpenUI: player_event_on_command = (event: number,player: Player, command: string): boolean => {
+    if(command == 'mythicplus') {            
+        const state = StateStorage.get(player.GetGUIDLow());
+        logger.debug(`OpenUI command 
+            player: ${player.GetName()}, 
+            difficulty ${state.difficulty}, 
+            groupId: ${state.groupId}, 
+            groupLeader: ${state.groupLeader}, 
+            isLeader: ${state.isLeader}`
+        );
+        aio.Handle(player, 'MythicPlus', 'ShowUI', StateStorage.get(player.GetGUIDLow())); 
         return false; 
     }
     return true; 
 }; 
 
+RegisterPlayerEvent(PlayerEvents.PLAYER_EVENT_ON_COMMAND, (...args) => OpenUI(...args));
 
+const MPStartState: player_event_on_login = (_event: number, player: Player): void => {
+    _refreshState(player);
+};
+
+// On login set up the mythic panel mod state for the player
+RegisterPlayerEvent(PlayerEvents.PLAYER_EVENT_ON_LOGIN, (...args) => MPStartState(...args));
+
+// API Handlers available to the client
 const MPHandlers = aio.AddHandlers("MythicPlus", {    
-    GetDifficulty, 
-    SetDifficulty
+    SetDifficulty, 
+    GetState,
 }); 
-
-RegisterPlayerEvent(PlayerEvents.PLAYER_EVENT_ON_COMMAND, (...args) => ShowIt(...args));
